Estimate Exponential-based model expmodel

Estimates Exponential-based model expmodel among parameter vectors, deglist, lmdlist. Then it sorts the results by AIC.

Usage

# S3 method for expmodel
estimate(
  model,
  deglist = deglist,
  lmdlist = lmdlist,
  recompute = FALSE,
  stepsize = NULL,
  verbose = FALSE,
  ...
)

Arguments

model

An object of expmodel class.

deglist

A vector of degrees of polynomials. They should be positive integers.

lmdlist

A vector of rate parameters of Exponential-based models. They should be larger than 0.

recompute

If TRUE, recomputes the results for better estimation and accuracy. Parameters whose accuracies had been already attained sufficiently, namely around 1.0e-6, are not included in candidates for recomputing.

stepsize

A vector in descending order whose values are between 0 and

If a small step size is supplied, it can attain successful estimates, but it might take more iterations.

verbose

If TRUE, it shows the detailed message of SDP solver.

...

Arguments to be passed to or from other methods.

Value

A expmodel object including the estimates. Those estimates are stored in model$result with data.frame

form and model$coeffs in list form.

See also

Examples

## Create an expmodel object
emodel <- expmodel(mixexpgamma$n200)
## Estimate a model with parameters
emodel <- estimate(emodel, c(4,5,6), c(0.25, 0.5, 1, 2, 4))
#> (4,0.25)
#> Status: Stagnation1: Rollback
#> Status: Stagnation
#> (4,0.5)
#> Status: Normal Termination.
#> (4,1)
#> Status: Normal Termination.
#> (4,2)
#> Status: Normal Termination.
#> (4,4)
#> Status: Normal Termination.
#> (5,0.25)
#> Status: Normal Termination.
#> (5,0.5)
#> Status: Stagnation1: Rollback
#> Status: Normal Termination.
#> (5,1)
#> Status: Stagnation2: Rollback
#> Status: Normal Termination.
#> (5,2)
#> Status: Normal Termination.
#> (5,4)
#> Status: Normal Termination.
#> (6,0.25)
#> Status: Stagnation1: Rollback
#> Status: Stagnation
#> (6,0.5)
#> Status: Stagnation1: Rollback
#> Status: Stagnation
#> (6,1)
#> Status: Stagnation2: Rollback
#> Status: Stagnation2
#> (6,2)
#> Status: Normal Termination.
#> (6,4)
#> Status: Stagnation1: Rollback
#> Status: Normal Termination.