Estimate Gaussian-based model gaussmodel

Estimates Gaussian-based model gaussmodel among parameter vectors, deglist, mulist, sdlist. Then it sorts the results by AIC.

Usage

# S3 method for gaussmodel
estimate(
  model,
  deglist = deglist,
  mulist = mulist,
  sdlist = sdlist,
  scaling = FALSE,
  recompute = FALSE,
  stepsize = NULL,
  verbose = FALSE,
  ...
)

Arguments

model

An object of a gaussmodel class.

deglist

A vector of degrees of polynomials. They should be positive even numbers.

mulist

A vector of means for Gaussian-based models.

sdlist

A vector of standard deviations for Gaussian-based models. They should be larger than 0.

scaling

A logical scalar, which indicates whether or not it scales means and standard deviations in mulist and sdlist. The default value is FALSE.

recompute

If TRUE, recomputes the results for better estimation and accuracy. Parameters whose accuracies had been already attained sufficiently, namely around 1.0e-6, are not included in candidates for recomputing.

stepsize

A vector in descending order whose values are between 0 and

If a small step size is supplied, it can attain successful estimates, but it might take more iterations.

verbose

If TRUE, it shows the detailed message of SDP solver.

...

Arguments to be passed to or from other methods.

Value

A gaussmodel object including the estimates. Those estimates are stored in model$result with data.frame

form and model$coeffs in list form.

See also

Examples

## Create an `gaussmodel` object
gmodel <- gaussmodel(mix2gauss$n200)
## Estimate a model with parameters
gmodel <- estimate(gmodel, deglist=c(2, 4), mulist=c(0.0, 0.2),
             sdlist=c(0.75, 1.0, 1.25))
#> (2,0,0.75)
#> Status: Normal Termination.
#> (2,0,1)
#> Status: Normal Termination.
#> (2,0,1.25)
#> Status: Normal Termination.
#> (2,0.2,0.75)
#> Status: Normal Termination.
#> (2,0.2,1)
#> Status: Normal Termination.
#> (2,0.2,1.25)
#> Status: Normal Termination.
#> (4,0,0.75)
#> Status: Normal Termination.
#> (4,0,1)
#> Status: Normal Termination.
#> (4,0,1.25)
#> Status: Normal Termination.
#> (4,0.2,0.75)
#> Status: Normal Termination.
#> (4,0.2,1)
#> Status: Normal Termination.
#> (4,0.2,1.25)
#> Status: Normal Termination.