Constructor for S3 class expmodel

This function is a constructor for S3 class expmodel, which represents Exponential-based model. It usually takes data and optionally freq as arguments and also optionally stepsize Members of interest in practice are result and coeffs, which maintain the information of estimates and coefficients of polynomials, respectively.

Usage

expmodel(data = data, freq = NULL, stepsize = c(0.5, 0.3))

Arguments

data

A nonnegative numeric vector of a data set to be estimated.

freq

A frequency vector corresponding to the data vector. The default value is NULL, which means all frequencies are one. If supplied, the length of a vector should be same as data and each element should be a nonnegative integer.

stepsize

A numeric vector whose element is larger than 0 and smaller than 1, and decreasing order. The default value is c(0.5, 0.3). If you encounter numerical difficulties, decreasing its values, for example, to c(0.4, 0.2), might help to estimate a model.

Value

An object of Exponential-based model expmodel.

See also

Examples

## Create `expmodel` object from a data set `mixexpgamma$n200`.
 emodel <- expmodel(mixexpgamma$n200)
## Create `expmodel` object from a data set `mixExpGammaHist$n800p` and
## its frequencies `mixExpGammaHist$n800f`.
 emodel <- expmodel(mixExpGammaHist$n800p, mixExpGammaHist$n800f)